# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio API",
        "description": "Backend API for the portfolio project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{scheme}://{host}",
            "description": "Dynamic server",
            "variables": {
                "scheme": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "default": "localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/api/v1/payments/intent": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create payment intent",
                "operationId": "244df0a5fe6f025b0ca856191a233c28",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_id"
                                ],
                                "properties": {
                                    "booking_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payments/confirm": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Confirm payment",
                "operationId": "44fe27737fbdf48934f11b59ace7a4b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_id",
                                    "payment_intent_id",
                                    "payment_method"
                                ],
                                "properties": {
                                    "booking_id": {
                                        "type": "integer"
                                    },
                                    "payment_intent_id": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "stripe",
                                            "card",
                                            "online"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payments/webhook": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Handle Stripe webhook",
                "operationId": "4ce4817727fef48fb4e50d0d0555595d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiError": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "ValidationError"
                    },
                    "code": {
                        "type": "string",
                        "example": "VALIDATION_FAILED"
                    },
                    "details": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "OpenApi",
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "ApiMeta": {
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "page_size": {
                        "type": "integer",
                        "example": 15
                    },
                    "total_count": {
                        "type": "integer",
                        "example": 120
                    },
                    "total_pages": {
                        "type": "integer",
                        "example": 8
                    },
                    "has_next_page": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_previous_page": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "ApiEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {},
                    "error": {
                        "$ref": "#/components/schemas/ApiError"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/ApiMeta"
                    },
                    "trace_id": {
                        "type": "string",
                        "example": "c5d5f3fa-5a7f-4c7a-9c7f-0b2a6e3f2f1e"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Send as: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}